feat: encapsulate 3D data augmentation
